consolidate pull and push to http
