+++ internal/controller/discovery/loader.go
@@ -5,7 +5,7 @@ import (
 
 	gnmicv1alpha1 "github.com/gnmic/operator/api/v1alpha1"
 	"github.com/gnmic/operator/internal/controller/discovery/core"
-	pull "github.com/gnmic/operator/internal/controller/discovery/loaders/pull"
+	http "github.com/gnmic/operator/internal/controller/discovery/loaders/http"
 )
 
 // NewLoader creates a loader by name
@@ -14,7 +14,7 @@ func NewLoader(name string, namespace string, spec gnmicv1alpha1.TargetSourceSpe
 
 	switch {
 	case spec.Provider.HTTP != nil:
-		return pull.New(cfg), nil
+		return http.New(cfg), nil
 	case spec.Provider.Consul != nil:
 		return nil, fmt.Errorf("unknown targetsource loader, check TargetSource CRD for %s", loaderName)
 	default:
