+++ internal/controller/discovery/loaders/all/all.go
@@ -1,6 +1,5 @@
 package all
 
 import (
-	_ "github.com/gnmic/operator/internal/controller/discovery/loaders/pull"
-	// _ "github.com/gnmic/operator/internal/controller/targetsource/loaders/push"
+	_ "github.com/gnmic/operator/internal/controller/discovery/loaders/http"
 )
