+++ internal/controller/discovery/loaders/http/loader_test.go
@@ -1,0 +1,1 @@
+package http
