--- internal/controller/discovery/loaders/pull/loader_test.go
@@ -1,1 +1,0 @@
-package pull
