--- internal/controller/discovery/loaders/push/loader_test.go
@@ -1,1 +1,0 @@
-package push
